jwt/requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==3.2.11
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.13.1

django-cors-headers==3.5.0
psycopg2-binary==2.8.5
pandas==1.1.1
scikit-learn==0.23.2
joblib==0.16.0
gunicorn==20.0.4
Pillow==9.0.1

pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
